requirements.txt
fastapi
uvicorn
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
alembic

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
factory-boy==3.3.0
